feat(V3.5-Phase1): 内存池模块硬件测试完成 + 文档更新
✅ 完成功能:
- 内存池基础功能 (分配/释放)
- 满池保护机制
- 泄漏检测 (6秒自动告警 + 持续监控)
- 双重释放检测
- 峰值使用统计
- 并发安全保证

✅ 测试结果:
- 压力测试: 200次操作, 0失败, <1μs/op
- 泄漏检测: 6秒触发首次告警, 后续每秒持续告警
- 并发测试: 满池保护 + 双重释放检测通过
- 帧池峰值: 4/4 (100%)
- 状态池峰值: 8/8 (100%)

📄 文档更新:
- V3.5-Phase1-内存池测试指南.md
  - 修正泄漏检测流程 (自动持续告警)
  - 添加实际测试结果和数据
  - 更新验收标准

🔧 关键修复:
- 禁用 Modbus Gateway 轮询 (FEATURE_MODBUS_ENABLE=0)
- 修复 IDLE 中断处理 (先保存SR, 再读DR)
- 验证 RS485 通信正常 (电机响应测试通过)
- 注释未实现的调试函数 (避免链接错误)

📊 资源占用:
- Flash: 43572B (66.49%)
- RAM: 7336B (35.82%)

✅ 可投入生产使用

Next: V3.5 Phase 2 - 多电机管理优化

+++ Drivers/BSP/MEM_POOL/mem_pool.c
@@ -180,31 +180,31 @@ void mem_pool_print_stats(void)
     printf("[Frame Buffer Pool] (%d x %dB = %dB total)\r\n", 
            MEM_POOL_FRAME_COUNT, MEM_POOL_FRAME_SIZE,
            MEM_POOL_FRAME_COUNT * MEM_POOL_FRAME_SIZE);
-    printf("  Total Allocs:      %lu\r\n", s_frame_ctrl.stats.total_allocs);
-    printf("  Total Frees:       %lu\r\n", s_frame_ctrl.stats.total_frees);
-    printf("  Current Used:      %lu / %d blocks (%d%%)\r\n", 
-           s_frame_ctrl.stats.current_used, MEM_POOL_FRAME_COUNT,
-           (s_frame_ctrl.stats.current_used * 100) / MEM_POOL_FRAME_COUNT);
-    printf("  Peak Used:         %lu blocks\r\n", s_frame_ctrl.stats.peak_used);
-    printf("  Alloc Failures:    %lu\r\n", s_frame_ctrl.stats.alloc_failures);
-    printf("  Invalid Frees:     %lu\r\n", s_frame_ctrl.stats.invalid_frees);
-    printf("  Double Frees:      %lu\r\n", s_frame_ctrl.stats.double_frees);
-    printf("  Leak Warnings:     %lu\r\n", s_frame_ctrl.stats.leak_warnings);
+    printf("  Total Allocs:      %lu\r\n", (unsigned long)s_frame_ctrl.stats.total_allocs);
+    printf("  Total Frees:       %lu\r\n", (unsigned long)s_frame_ctrl.stats.total_frees);
+    printf("  Current Used:      %lu / %d blocks (%lu%%)\r\n", 
+           (unsigned long)s_frame_ctrl.stats.current_used, MEM_POOL_FRAME_COUNT,
+           (unsigned long)(s_frame_ctrl.stats.current_used * 100) / MEM_POOL_FRAME_COUNT);
+    printf("  Peak Used:         %lu blocks\r\n", (unsigned long)s_frame_ctrl.stats.peak_used);
+    printf("  Alloc Failures:    %lu\r\n", (unsigned long)s_frame_ctrl.stats.alloc_failures);
+    printf("  Invalid Frees:     %lu\r\n", (unsigned long)s_frame_ctrl.stats.invalid_frees);
+    printf("  Double Frees:      %lu\r\n", (unsigned long)s_frame_ctrl.stats.double_frees);
+    printf("  Leak Warnings:     %lu\r\n", (unsigned long)s_frame_ctrl.stats.leak_warnings);
     
     /* 电机状态池 */
     printf("\r\n[Motor State Pool] (%d x %dB = %dB total)\r\n", 
            MEM_POOL_MOTOR_STATE_COUNT, MEM_POOL_MOTOR_STATE_SIZE,
            MEM_POOL_MOTOR_STATE_COUNT * MEM_POOL_MOTOR_STATE_SIZE);
-    printf("  Total Allocs:      %lu\r\n", s_motor_state_ctrl.stats.total_allocs);
-    printf("  Total Frees:       %lu\r\n", s_motor_state_ctrl.stats.total_frees);
-    printf("  Current Used:      %lu / %d blocks (%d%%)\r\n", 
-           s_motor_state_ctrl.stats.current_used, MEM_POOL_MOTOR_STATE_COUNT,
-           (s_motor_state_ctrl.stats.current_used * 100) / MEM_POOL_MOTOR_STATE_COUNT);
-    printf("  Peak Used:         %lu blocks\r\n", s_motor_state_ctrl.stats.peak_used);
-    printf("  Alloc Failures:    %lu\r\n", s_motor_state_ctrl.stats.alloc_failures);
-    printf("  Invalid Frees:     %lu\r\n", s_motor_state_ctrl.stats.invalid_frees);
-    printf("  Double Frees:      %lu\r\n", s_motor_state_ctrl.stats.double_frees);
-    printf("  Leak Warnings:     %lu\r\n", s_motor_state_ctrl.stats.leak_warnings);
+    printf("  Total Allocs:      %lu\r\n", (unsigned long)s_motor_state_ctrl.stats.total_allocs);
+    printf("  Total Frees:       %lu\r\n", (unsigned long)s_motor_state_ctrl.stats.total_frees);
+    printf("  Current Used:      %lu / %d blocks (%lu%%)\r\n", 
+           (unsigned long)s_motor_state_ctrl.stats.current_used, MEM_POOL_MOTOR_STATE_COUNT,
+           (unsigned long)(s_motor_state_ctrl.stats.current_used * 100) / MEM_POOL_MOTOR_STATE_COUNT);
+    printf("  Peak Used:         %lu blocks\r\n", (unsigned long)s_motor_state_ctrl.stats.peak_used);
+    printf("  Alloc Failures:    %lu\r\n", (unsigned long)s_motor_state_ctrl.stats.alloc_failures);
+    printf("  Invalid Frees:     %lu\r\n", (unsigned long)s_motor_state_ctrl.stats.invalid_frees);
+    printf("  Double Frees:      %lu\r\n", (unsigned long)s_motor_state_ctrl.stats.double_frees);
+    printf("  Leak Warnings:     %lu\r\n", (unsigned long)s_motor_state_ctrl.stats.leak_warnings);
     
     printf("============================================\r\n\r\n");
 }
